refactor: consolidate IsSigningTransaction logic into Transaction type and remove redundant function
