+++ core/types/transaction_viction.go
@@ -1,6 +1,13 @@
 package types
 
-import "github.com/ethereum/go-ethereum/common"
+import (
+	"bytes"
+
+	"github.com/ethereum/go-ethereum/common"
+)
+
+// signMethodSelector is the 4-byte function selector for sign(uint256,bytes32).
+var signMethodSelector = common.Hex2Bytes("e341eaa4")
 
 // IsTomoXApplyTransaction returns true if the transaction is directed to the given TomoXContract address.
 func (tx *Transaction) IsTomoXApplyTransaction(contractAddress common.Address) bool {
@@ -15,15 +22,6 @@ func (tx *Transaction) IsSkipNonceTransaction(contractAddress common.Address) bo
 	return tx.IsTomoXApplyTransaction(contractAddress)
 }
 
-// IsSigningTransaction returns true if the transaction is a block signer registration tx.
-func (tx *Transaction) IsSigningTransaction() bool {
-	if tx.To() == nil {
-		return false
-	}
-	// Signer contract address is 0x88
-	return *tx.To() == common.HexToAddress("0x0000000000000000000000000000000000000088")
-}
-
 // IsTradingTransaction checks if a target address is one of the designated system exchange contracts
 func IsTradingTransaction(to *common.Address) bool {
 	if to == nil {
@@ -35,4 +33,28 @@ func IsTradingTransaction(to *common.Address) bool {
 		addr == common.HexToAddress("0x0000000000000000000000000000000000000092") || // TradingState
 		addr == common.HexToAddress("0x0000000000000000000000000000000000000093") || // Lending
 		addr == common.HexToAddress("0x0000000000000000000000000000000000000094") // LendingFinalized
+}
+
+// IsSigningTransaction returns true if the transaction is a block-signer
+// registration transaction to the BlockSigner contract.
+// blockSignAddr is the ValidatorBlockSignContract address from chain config.
+func (tx *Transaction) IsSigningTransaction(blockSignAddr common.Address) bool {
+	if tx.To() == nil {
+		return false
+	}
+	if *tx.To() != blockSignAddr {
+		return false
+	}
+	data := tx.Data()
+	if len(data) < 4 {
+		return false
+	}
+	if !bytes.Equal(data[0:4], signMethodSelector) {
+		return false
+	}
+	// sign(uint256 blockNumber, bytes32 blockHash) = 4 + 32 + 32 = 68 bytes
+	if len(data) != 68 {
+		return false
+	}
+	return true
 }
